add nex script to pick random queries + to generate plots

+++ scripts/run_workload.sh
@@ -2,6 +2,13 @@
 # Run the tpf client with a workload of queries
 
 WORKLOAD=$*
+
+if [ "$#" -ne 1 ]; then
+  echo "Illegal number of parameters."
+  echo "Usage: ./run_workload.sh <queries>"
+  exit
+fi
+
 rm -rf execution_times.csv results errors
 mkdir -p results/
 mkdir -p errors/
