+++ scripts/run_workload_file.sh
@@ -1,8 +1,15 @@
 #!/bin/bash
 # un the tpf client with a workload of queries, all stored in the same file
 
-cpt=1
 QUERIES=$1
+cpt=1
+
+if [ "$#" -ne 1 ]; then
+  echo "Illegal number of parameters."
+  echo "Usage: ./run_workload_file.sh <queries-file>"
+  exit
+fi
+
 rm -rf execution_times.csv queries results errors
 mkdir -p queries/
 mkdir -p results/
